+++ src/controllers/UserController.ts
@@ -1,9 +1,10 @@
 import { Request, Response } from "express";
-import { getCustomRepository } from "typeorm";
-import { UsersRepository } from "../repositories/UsersRepository";
+
 import * as yup from "yup";
 import { AppError } from "../errors/AppError";
+import { UsersServices } from "../services/UsersServices";
 
+const usersServices = new UsersServices();
 class UserController {
   async create(request: Request, response: Response) {
     const { name, email } = request.body;
@@ -19,22 +20,15 @@ class UserController {
       throw new AppError(error);
     }
 
-    const usersRepository = getCustomRepository(UsersRepository);
-    const userAlreadExists = await usersRepository.findOne({
-      email,
-    });
+    try {
+      const user = await usersServices.create(name, email);
 
-    if (userAlreadExists) {
-      throw new AppError("User alread exists!");
+      return response.status(201).json(user);
+    } catch (error) {
+      console.log(">>>>>>>>>>>>>", error);
+
+      throw new AppError(error);
     }
-    const user = usersRepository.create({
-      name,
-      email,
-    });
-
-    await usersRepository.save(user);
-
-    return response.status(201).json(user);
   }
 }
